fixed operator script parameter translation bug

+++ trunk/XStudio2/Logic/ScriptObjectLibrary.h
@@ -60,7 +60,7 @@ namespace Logic
                   return it->second;
 
                // Error: Not found
-               throw ScriptObjectNotFoundException(HERE, (UINT)grp, id);
+               throw ScriptObjectNotFoundException(HERE, grp, id);
             }
 
             // ----------------------- MUTATORS ------------------------
